indent/nix: remove equals sign logic
This commit fixes double-indent problems. Previously, any code with a
bracket on the next line would be indented like this:

{
  foo =
    {
      # todo
    };
}

After this commit, it's now properly indented like this:

{
  foo =
  {
    # todo
  };
}

This does cause some slight regressions, specifically when an equals
sign starts an expression that should be indented over. For example,
this code:

{
  combinedList =
    list1
    ++ list2
    ++ list3;
}

Now regresses into this:

{
  combinedList =
  list1
  ++ list2
  ++ list3;
}

To prevent these regressions, a multiline regex should really be used,
which would only increase indentation on an equals sign if:

1. the current line doesn't contain a semicolon
2. the next line didn't contain a starting bracket.

This seems quite simple in theory, I just wasn't able to get it working
myself. Some help on implementing this would be great if we'd prefer to
fix these regressions before merging.